requirements.txt
streamlit
vanna
psycopg[binary]
python-dotenv
openai
chromadb
pandas
pysqlite3-binary
protobuf==3.20.3
psycopg2-binary>=2.9.10
openpyxl
databricks-sql-connector

# vanna #[chromadb,openai]
# streamlit
# python-dotenv
# chromadb
# openai
# vanna[postgres]
# psycopg[binary]


# psycopg2-binary
# pandas
# requests
# datetime